Deleted skeletons and created project structure
It seemed like the skeletons where a good place to start with the
Arrowhead code, but the SOS examples have more features that will be
needed so will use them as the base. The basic demo-car will be a good
starting point.

Using the basic mvn archetype quickstart and copying the resources file
I have created a project with the same structure as the one used for
the Arrowhead systems. This one is where the code from the SOS will be
merged to provide the Arrowhead envelope for the Workflow Executor.