Added first classes to Workflow Executor
Using the same structure as in the demo-car I have created a constant
class and a Spring boot application.

The Spring start the application and the code run at init instead of
being written in main, is written in the ApplicationListener at start
and at the end.

For the moment the code only tries to contact the core system with the
configuration from the properties files, and register two services.

The certificates have not been modified yet, I will need to create my
own certificates.