Added Rest Controller class to Workflow Executor
The Controller is the one in charge of anwsering the services request
given by the Spring dispatcher. At the moment only implemented a simple
echo service, to chek the state of the System.

Found problem when running the system, the embedded Tomcat Server has
an Error at the end of Application Init, not sure why is caused. The
property spring.main.web-application-type is none by default, but then
no server is initialized, and when option changed to servlet, the error
is thrown. Next steps will be debugging until the problem is found.