Fixed error with server and added logging code
The error that prevented the server correct initialization was due to
a problem in the certificate, the alias field had a typo. Maybe next
steps could be to make my own certificate.

In the process I added logging information when registering/deleting
the services in the Service Registry, and some test trying to access
the emebedded server. However the server tests seemed to access another
server, as the ports did not match.