Fixed errors server: not auto start and shut down
Somehow the embedded server (tomcat) was only initialized when the
property web-application-type was explicitly set to "servlet", but in
producer skeletons and examples this was not necessary. Problem was
with spring boot not finding the web application context and embedded
server in the dependencies.

The web-application-type set to "none" causes the application to shut
down as soon as the init code is done, shut down after start up. Since
it does not initiate a server, the application will unregister its
services and stop. Therefore the property should be set to "servlet" or
commented for simplicity.