Add Emanuel scripts to create cerficates
New folder with the scripts taken from the core arrowhead repository
and the example modified to be used for the workflow executor.

By default the script will create certificates p12 and pub with the
word "test" in the name. If we want to change the name we have to
modify script, it takes no command line arguments. Also if we need
to change the root cert or cloud, we need to modify script and add the
new files to resource folder.

Created a folder with the necessary scripts to create all the
certificates