Add waiting method to test core system is ready
Create a new helping method that waits for a period of time in minutes
during which makes repeatedly request to core system to check its
status.

As this system depends on the other Core systems, if at the end of the
method the core systems are still down, it will fails fast gracefully
by throwing an exception.