Add StateMachine project and fix test dependency
Added the package of the project with the StateMachine logic to be used
by the Workflow Execeutor. This brought problems with the test classes
as before it was a standalone java project and not a maven project, so
it could not find the junit 5 dependencies. Therefore the pom file was
updated, while the versions of the junit 5 dependencies are managed by
Spring.