Add skeleton code for provided services
There is code in place that already registers the services that this
system will provide, but the services themselves are not yet
implemented. There was a typo in metadata that is now fixed.

First the services are created in the controller and the data that they
will receive and supply specified in the DTO classes.