Enhance DTO and created business object in service layer
The DTO name has been changed, according to the business object it
represents. The new QueuedWorkflowDTO stores information from a new
QueuedWorkflow business object.

The constructors for the DTOs have been written, although the
QueuedWorkflowDTO needs revision after last min change. The first idea
was to use Builder pattern, but then discarded as too complex for this
class and used a small set of static factory methods, but then also
found unuseful and left with only one method. Still incomplete.

The new Workflow business class wraps a State Machine with useful data.
It includes a name, configuration to be introduced into environment, a
StateMachine and a Workflow Status.

The Workflow Status sits in his own enum class, not inside any other.
It indicates if the workflow is under execution or not, or if it is
finished. This field was first introduced in the child class
QueuedWorkflow but it seemed to be used and defined by the parent, so
it was transferred.

The QueuedWorkflow business was renamed from the ActiveWorkflow class
and set to extend the basic Workflow class. It is unclear still what
the relation is, if a "is a" as a child class, or a "has a" as a field.

The QueuedWorkflow extends the data in basic Workflow so that each
object instance can be tracked and stored in the system. Therefore it
has a unique id, and 3 time fields: for workflow registration,
activation and end of execution.