Add service class to execute Workflow and finish init
Create new WExecutorService class that stores the different Workflows
in two data structures, a set for the different Workflow templates and
a Queue for the specific workflows that are waiting to be executed.

The goal is to have this class as a singleton bean to be called by the
controller upon receiving service request. It will also be used in the
initialization code, to add the set up of the Workflow types, before
the controller is online.

Modify the other service classes from previous commit, add missing
encapsulation by modified the access of fields to private.

Next step is to add behaviour to this classes.