Add methods to Service layer to be used by Controller
Created methods in WExecutorService to be called by the Rest controller
upon web requests. These methods include the basic operation that can
be done to the resource, well mapped with the services offered.

But methods in Service class are incomplete, lacking behaviour.

Controller uses the methods and does the conversion from Entities to
DTOs, instead of doing the conversion in the Service layer as the
arrowhead examples. To convert between objects it uses the mapping
methods of each DTO class.