Updated REST API of provided services
The URI of the resources has been modified trying to follow a more
consistent structure, and their descriptions for the Service Registry
has been updated accordingly.

The execute workflow service has been updated to obtain the workflow
information needed as input from the body of request, instead of the
request parameters.