Update initialization code of preloaded workflows
Create new class InitialWorkflows to contain the workflows needed by
the system at start up. Added to Spring as a bean. to be used as a
singleton in the WExecutor Service. At this point, there is only one
workflow under construction to test funcionalities.

Remove the code that used to load this workflows, in Application
Listener, as they are new injected by Spring.

Update the WExecutor service also, to include the init class as an
autowired dependency.

Add some missing parameters to the response of provided services in
Controller class.