Fix equals and hasCode methods of Workflow class
The equals method of Workflow had been changed to be able to search
and compare between the workflows. It was in a rough state, now it has
been polished.

Together with the equals method, we need to update the hashCode method
to maintain their contract. Used the default implementation provided by
Eclipse for the hashCode. It should be enough for this prototype, as is
mainly used for the HashSet that stores the types of Workflows known.