Update thread-safe class to store waiting Workflows
The implementation of Queue interface was changed to another thread-safe
class with operations that block the thread when certain conditions are
met. The previous implementation offers better performance when there is
simulatenously multiple consumers, or multiple producers, as it uses
CAS not locks to reduce contention. However the use case does not point
towards that scenario being the norm, otherwise the implementation
would be reverted to the previous class.