Review of ".service" package classes, refactoring
Encapsulate workflow starting process in startWorkflow method of
Workflow class, also used by QueuedWorkflow child class to add its
behavior. Simplified use in spun thread of WExecutorService, now the
process of executing a Workflow is just a method call.

Changed WorkflowConfig type, as it intended use match more that of a
list of Strings for the values of its entries, not just String.

Detected problem when executing workflows. As the QueuedWorkflow is not
a copy of the Workflow Object, but the original, changes are being
propagated. That is why the execution of a Workflow only happens the
first time it is invoked, and idles the rest.

Review Workflow Status Javadoc