Fix bug in State Machine execution & add object copy
Found bug in StateMachine class. Its update method always returned true
,instead of true or false depending if we reach the end of the State
Machine.

Found missing functionality to copy objects in Workflow and State
Machine classes. If the Workflows loaded at init time are to be use as
templates, there is a need to create objects from them during runtime.
Added copy constructors to both clases, as the use of the clone()
methos was not advised. Also added some getters and setters that were
found missing for these clases.
This updates pushed changes to the WExecutorService, that searches in
the templates and queues the Workflows.

Update the startWorkflow procedure. In previous commit it has been
refactored, but it showed new issues. Now is more concise, but there is
still a issue. The return values of these methods need to be clarified.
This affected the Thread executing the workflows, in WExecutorService,
that had not been updated after the refactor. It includes more
information now, as timestamps of when the Workflows started/finished.