Fix log4j2 xml file used as configuration and format
The log4j2 is using to configure the logs the "log4j2.xml" present in
target directory, not inside jar. In this way the logging can change
without the need to recompile project.

Also change the pattern format for the log files created. Its pattern
is esentially the same, but without the color options, as those are for
displaying nicely in console only.