Redesign of InitialWorkflows and add log to workflows
First the InitialWorkflows class went through tests in the actual state
of the Workflow Executor, and it threw exceptions due to its own code.
There were errors in code, and a bug in the commons Arrowhead library
that was reported in github. Most of them are fixed now.

As the test were being done, it became clear that better tools for
debugging were needed. So logs were added to the Workflow class, plus
the ones that will be inserted in the test State Machine still in
progress.

Finally Javadocs for the InitialWorkflows were written, perhaps need to
review some of them or add to the private methods. This helped
clarifying the purpose of the State Machine and its behavior.

TODO is the implementation according to the Javadocs, to match comments
and code.