Fix bugs, like WorkflowConfig keeping default values
Fix small bugs found after testing code from last big commit.

The InitialWorkflows had a bug in new method printCurrentState(String),
that cause to always throw an exception. It was due to missing
return statement, and the exception message was also updated to provide
better information.

The WExecutorApplicationListener had a bug difficult to catch. If the
WorkflowExecutor system was started before Orchestrator was finished
with its start-up process, it will never call the method
updateCoreServiceURIs(), and the error will be thrown at a later time
in the State Machine when calling for orchestration. Now the call to
the updateCoreServiceURIs() is made after Orchestrator responds to echo
request.

In the WExecutorService class, when the Workflows where prepared to be
included in the execution queue, they were first updated with the
configuration provided when calling the service. But this process only
overwrites the keys included in the service call, and so the type
values present in the original Workflow will stay if they were not
updated in the request. This behavior is not wanted, now the
configuration is completely wiped before the values in the request are
introduced.