Rewrite InitialWorkflows class and test it
The InitialWorkflows class has suffered a major change in structure.
The main idea of the testWorkflowExecutor State Machine is the same as
in previous versions, to test the services offered by the Workflow
Executor system. Although, as stated in last commit through its Javadoc
, now it has less states and more transitions, and the implementation
reflects the Javadoc documentation.

There are only 3 states, but 5 transitions. State 1 has transitions 1,
2, and 3 going out and back into State 1. In each one of this
transitions a different service from the Workflow Executor is tested.
Then the last transition 4 goes from State 1 into 2 to finish the
State Machine.

After the code was completed it was tested, and there were some errors
found:

- InitialWorkflow had a printCurrentState method which worked
incorrectly, as it always showed the same result. It was accessing the
Workflow type object, not the QueuedWorkflow under execution.

- QueuedWorkflowDTO required for object creation a valid QueuedWorkflow
but it was used as input to a method to retrieve its class, not its
state. To solve the situation QueuedWorkflowDTO now contains a default
constructor, with its fields to null or invalid values. Perhaps this
could change in the future.

- WorkflowDTO and WExecutorService used at some point the Workflow
Configuration, but did not check for null. As it is an optional
parameter a null value is accepted, so it needs to perform the null
tests internally-

- WExecutorService had a bug in the "workflow consumer" thread. When
retrieving a Workflow from the Queue, it was removing it as that was
required by the thread blocking methods. This caused an issue for the
service "provide-workflows-in-execution" as it always returned null.
Now the workflows are retrieved as before, to block the thread when
Queue is empty, but they are added inmediately after, and removed after
execution is complete.