Review and clean code and comments
Remove useless code in InitialWorkflows class. Mainly related to the
method for obtaining the currentState, as it is not feasible at the
moment. A possible implementation needs injection of  objects that
should not be accessed by a StateMachine, as they are internal to the
Workflow Executor system.

Updated echo service description, as described in a github issue of the
core systems.

Removed TODO comments in Workflow and StateMachine class, as the issues
are now resolved or a btter solutions was not found.