Change return value while updating StateMachine
Change the return value of the update() method in StateMachine class in
charge of executing one step of the State Machine/Workflow. Before, the
returned value was a bool, signaling with a false value the end state
of the State Machine. Now it returns a wrapper object that contains
more information, like how was the update of the State Machine (was
a transition triggered?), in addition to also signaling the end state.
The object also includes a reference to the state and transition, this
way the user of the State Machine can know which of the many
transitions was taken to arrive to that state.

This change made necessary to also change the Workflow class, in charge
of executing the State Machine by calling the update() method. The
extra information returned is used to sleep the thread, when waiting
for changes from the outside to affect the State Machine.

It was also necessary to change the State, adding equals and hashCode
methods, in order to be able to compare the results obtained from
updating the State Machine.