Review code in WexecutorService and StateMachine
Found code in WexecutorService that was out of place, by being too
specific of a concrete StateMachine, commented it out and to be removed
in the future.

Found possible bug in StateMachine code, as setter method was not
checking the input for problems, like number out of range.
This also led to a better check in the constructor of the object, as
now there is no room for empty StateMachines without states or
transitions