Add error transition (17) to first State (0)
Add error transition (17) to first State (0), in the future to most of
the states. If the error Event is triggered, the State Machine will
jump to the state instead of continuing execution.

Also wrapped the functionality of the transitions (untill now only the
first one) in methods, so that they can be called from other State
Machines if they behave similarly.