Continue developing Fischer factory and successful first test
Implemented states and transitions to be able to execute a milling
operation, basically half way through the factory.

To test the FischeWorkflow class with the InitialWorkflow loading it,
it was necessary to Autowire it in InitialWorkflows and create a
PostConstruct method that will add the workflows in FischerWorkflow to
the set of workflows loaded at init of the WExecutor system.

Fix bug in new FischeWorkflow#consumeService() as the method for the
request (GET,PUT,...) was still being taken from the service metadata
instead of from the argument input.