Finish milling workflow but results are not returned
Finish and test the milling workflow in the Fischer factory.
All transitions and states work as expected, but some sensors had strict
timing requirements to pull their values and therefore the cycle time
of the WExecutor system had to be decreased to 2 milis.

Completed all transitions and events, although the correspondence from
state to transition need some polishing.

Still lacking is the logic to return the results of the State Machine
to a WManager system by consuming his services. This will be probably
done inside the WExecutorService class.