Modify Logic* classes to validate operator at construction
Split the logic inside the LogicOperator#evaluateOperator().
Now this method only evaluates the expression to true or false, it does
not test wether the inputs are adequate for that LogicOperator. A new
method was created to test if the LogicOperator can handle that number
of inputs.

The new method LogicOperator#testValidOperator is called when creating a
LogicExpression and it will throw a RuntimeException if the test fails.
The exception can not be handled, the code has to be changed, so it
should not be a checked exception.