Update WExecutor to send workflow results to WManager
Modify Workflow and QueuedWorkflow class with a revision of the code to
end a workflow. Changed what environment variables of State Machine
where reporting a succes or error, and also changed boolean return types
as they were not needed if WExecutorService can handle the Workflow
object.

Add funcitonality to WExecutorService to contact the WManager and
consume the result services that will be provided. For that created a
new class with useful methods that were already being used inside the
State Machines and modified the FinishWorkflowDTO to provide a class
reference, an object without valid state.

Created new WexecutorUtil class to store methods used in InitialWorkflows
and FischerWorkflows to consume Arrowhead services. Used another module
of client library to provide the beans needed to use those methods,
without being the class using them a bean itself.